face_detection/requirements.txt
datasets
deepface
Flask
numpy
Pillow
scipy
tf-keras
